requirements.txt
#VIT packages
numpy
matplotlib
torch
torchvision
timm
tqdm
sklearn
scikit-image
pandas
Pillow
wandb

#NLP packages
transformers

umap-learn
seaborn
